move bmi3d preproc code into submodule
