+++ tests/test_preproc.py
@@ -1,4 +1,5 @@
 from aopy.preproc import *
+from aopy.preproc.bmi3d import *
 from aopy.data import *
 import numpy as np
 import unittest
